Add backfilling of availability on event edit
If users already filled out availability on an event, and the date/time range changes when editing, this now updates the availabilities accordingly.

If timeslots are removed, the foreign key cascade will take care of that. If timeslots are added, then new availability entries will be created for each participant, defaulting to unavailable.